+++ tests/enzyme/test_vault_controlled_wallet.py
@@ -256,7 +256,8 @@ def test_vault_controlled_wallet_make_unauthorised(
     # Buy in to the vault
     usdc.functions.transfer(user_1, swap_amount).transact({"from": deployer})
     usdc.functions.approve(vault.comptroller.address, swap_amount).transact({"from": user_1})
-    vault.comptroller.functions.buyShares(swap_amount, 1).transact({"from": user_1})
+    tx_hash = vault.comptroller.functions.buyShares(swap_amount, 1).transact({"from": user_1})
+    assert_transaction_success_with_explanation(web3, tx_hash)
     assert usdc.functions.balanceOf(vault.address).call() == swap_amount
 
     # First approve tokens from the vault